first

+++ kill_sample_switch.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
 PIDS=`pgrep bf_switchd`
-if [ -n "$PIDS" ] # not empty
+MATCH=`ps aux | grep -E "$PIDS.{6,}bf_switchd.+sample_switch"`
+
+if [ -n "$PIDS" ] && [ -n "$MATCH" ] # not empty
 then
 	echo "These running switch processes will be killed: "$PIDS
 	kill -9 $PIDS > /dev/null 2>&1 #This will cause an error because it also kill the process of "grep", however it has terminated.
